+++ inform/__init__.py
@@ -4,7 +4,7 @@ from .inform import (
     Color, Info, LoggingCache,
 
     # user utility functions and classes
-    columns, conjoin, dedent, did_you_mean, fmt, format_range, full_stop,
+    bar, columns, conjoin, dedent, did_you_mean, fmt, format_range, full_stop,
     os_error, parse_range, plural, ProgressBar, render, render_bar, title_case,
     tree, truth,