+++ doc/releases.rst
@@ -9,6 +9,10 @@ Latest development release
     | Version: 1.34
     | Released: 2025-05-10
 
+- Added :class:`bar`.
+- The width of the *prefix* is now included in the width of the :class:`ProgressBar`.
+
+
 1.34 (2025-05-10)
 -----------------